+++ magemaker/azure/test_delete_model_integration.py
@@ -52,3 +52,23 @@ def test_delete_model():
         print("2. Check endpoint exists in Azure ML workspace")
         print("3. Verify Azure permissions")
         raise
+
+
+def test_deploy_huggingface_model_to_azure():
+
+    deployment = Deployment(
+        endpoint_name="test-endpoint",
+        instance_type="Standard_DS3_v2",
+        destination="azure",
+        instance_count=1,
+        quantization=None,
+        num_gpus=None
+    )
+
+    model = Model(
+        id="facebook-opt-125m",
+        source="huggingface",
+    )
+
+    deploy_huggingface_model_to_azure(deployment=deployment, model=model)
+    pass
