+++ magemaker/azure/test_delete_model_unit.py
@@ -3,6 +3,12 @@ from unittest.mock import patch, MagicMock
 from azure.core.exceptions import ResourceNotFoundError
 from magemaker.azure.delete_model import delete_azure_model
 
+from magemaker.schemas.model import Model
+from magemaker.schemas.deployment import Deployment
+
+from magemaker.azure.create_model import deploy_huggingface_model_to_azure
+
+
 @pytest.fixture
 def mock_env_vars():
     return {
@@ -11,6 +17,7 @@ def mock_env_vars():
         "AZURE_WORKSPACE_NAME": "test-workspace"
     }
 
+@pytest.mark.unit
 def test_delete_model_success(mock_env_vars):
     with patch('magemaker.azure.delete_model.dotenv_values') as mock_dotenv, \
          patch('magemaker.azure.delete_model.DefaultAzureCredential'), \
@@ -24,6 +31,7 @@ def test_delete_model_success(mock_env_vars):
         assert result is True
         mock_client.online_endpoints.begin_delete.assert_called_once()
 
+@pytest.mark.unit
 def test_delete_nonexistent_model(mock_env_vars):
     with patch('magemaker.azure.delete_model.dotenv_values') as mock_dotenv, \
          patch('magemaker.azure.delete_model.DefaultAzureCredential'), \
