+++ magemaker/sagemaker/test_create_model_unit.py
@@ -11,27 +11,29 @@ from magemaker.schemas.deployment import Deployment
 from magemaker.schemas.model import Model, ModelSource
 
 @pytest.fixture
-def sample_huggingface_model():
-    return Model(id="google-bert/bert-base-uncased", source=ModelSource.HuggingFace)
-
-@pytest.fixture
 def sample_deployment():
     return Deployment(destination="aws", instance_type="ml.m5.xlarge", instance_count=1)
 
 @pytest.mark.unit
 @patch('magemaker.sagemaker.create_model.S3Uploader.upload')
 @patch('magemaker.sagemaker.create_model.HuggingFaceModel')
-def test_custom_model_deployment(mock_hf_model, mock_s3_upload, sample_deployment, tmp_path):
+@patch('magemaker.sagemaker.create_model.sagemaker_session')
+def test_custom_model_deployment(sagemaker_session, mock_hf_model, mock_s3_upload, sample_deployment, tmp_path):
     # Create a mock model file
     test_model_file = tmp_path / "model.pt"
     test_model_file.write_text("dummy model content")
 
     # Mock S3 upload and model deployment
     mock_s3_upload.return_value = "s3://test-bucket/models/test-custom-model"
+
     mock_predictor = MagicMock()
     mock_predictor.endpoint_name = "test-endpoint-001"
+
     mock_hf_model_return = mock_hf_model.return_value
     mock_hf_model_return.deploy.return_value = mock_predictor
+
+
+    sagemaker_session.default_bucket.return_value = "test-bucket"
 
     custom_model = Model(
         id="test-custom-model",
