+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 markers =
     integration: marks tests as integration tests (deselect with '-m "not integration"')
+    unit: marks tests as unit tests
